Changes binding for PIDMOTD to PIDEndOfMOTD. We want to fire the handler only once, not on every MOTD line.
Implements proto.Client.Recover(). This allows recovering of our
nickname by re-authenticating it with our nickserv password.

Implements handling of PIDNickInUse message.
We try re-authenticating if we have a nickserv password. Otherwise
we change our nickname by appending an underscore to it and then
notify the server of the nick change.