Implements usage of a deep copies of a command when it is executed.
The command (and parameter) instances, created during plugin
initialization, should function as blueprints. When a command is
executed, current parameter values are assigned to the parameter
structures. Doing this on the originals, creates race condition
problems when multiple people issue the same command at the same time.

Working on a deep copy, solves this issue.